requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
black==21.4b2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
Django==3.2
django-debug-toolbar==3.2.1
django-ninja==0.13.0
django-ratelimit==3.0.1
django-seed==0.2.2
django-user-agents==0.4.0
djangorestframework==3.12.4
Faker==8.4.0
geoip2==4.2.0
gunicorn==20.1.0
idna==3.2
maxminddb==2.0.3
multidict==5.1.0
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.8.1
pydantic==1.8.2
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
tzdata==2021.5
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
yarl==1.6.3
